replaces vec with set during testing
this is mainly because the negative test had a loophole, where we
try to check if the uppercase version of a string is in the filter,
to assert a false positive.

for short lengths of strings, there is a good chance that this
will already be in the filter. so we use a set, and if it looks like
there is a false positive, we cross check if the word was not in the
set.

this gives an accurate response for false positives (currently 0)